b2b-doc-management/backend/requirements.txt
fastapi==0.95.2
uvicorn[standard]==0.23.2
python-multipart==0.0.6
Werkzeug==3.0.3

# Database and ORM
SQLAlchemy==1.4.48
mysqlclient==2.2.4

# Data handling
pandas==2.2.2
numpy==1.26.4

# Utilities
python-dotenv==1.0.1
PyJWT==2.8.0


